+++ nemdata/mmsdm.py
@@ -184,14 +184,23 @@ def download_mmsdm(
     for mmsdm_file in files:
         clean_fi = mmsdm_file.data_directory / "clean.parquet"
         if clean_fi.exists():
-            print(f" [blue]EXISTS[/] {' '.join(clean_fi.parts[-5:])}")
+            print(f" [blue]CACHED[/] {' '.join(clean_fi.parts[-5:])}")
             data = pd.read_parquet(clean_fi)
         else:
-            print(f" [blue]MISSING[/] {' '.join(clean_fi.parts[-5:])}")
+            print(f" [blue]NOT CACHED[/] {' '.join(clean_fi.parts[-5:])}")
+
+        data_available = utils.download_zipfile(mmsdm_file)
+
+        if not data_available:
+            print(
+                f" [red]NOT AVAILABLE[/] {' '.join(mmsdm_file.zipfile_path.parts[-5:])}"
+            )
+            data = None
+
+        else:
             print(
                 f" [green]DOWNLOADING[/] {' '.join(mmsdm_file.zipfile_path.parts[-5:])}"
             )
-            utils.download_zipfile(mmsdm_file)
             utils.unzip(mmsdm_file.zipfile_path)
             data = load_unzipped_mmsdm_file(mmsdm_file)
             assert table.datetime_columns
@@ -202,5 +211,11 @@ def download_mmsdm(
                 print(f" [green]SAVING [/] {clean_fi}")
                 data.to_csv(clean_fi.with_suffix(".csv"))
                 data.to_parquet(clean_fi.with_suffix(".parquet"))
-        dataset.append(data)
-    return pd.concat(dataset, axis=0)
+
+        if data is not None:
+            dataset.append(data)
+
+    try:
+        return pd.concat(dataset, axis=0)
+    except ValueError:
+        return pd.DataFrame()